Move the should send tests to the eth module
